requirements.txt
Django==1.5.5
argparse==1.2.1
distribute==0.6.34
django-picklefield==0.3.0
numpy==1.8.0
six==1.4.1
wsgiref==0.1.2
